+++ src/components/__tests__/FlawForm.spec.ts
@@ -99,7 +99,7 @@ describe('flawForm', () => {
     const subject = mountWithProps(flaw, { mode: 'edit' });
 
     const titleField = subject
-      .findAllComponents(LabelEditable)
+      .findAllComponents(LabelDiv)
       .find(component => component.props().label === 'Title');
     expect(titleField?.exists()).toBe(true);
 
@@ -182,7 +182,7 @@ describe('flawForm', () => {
     const subject = mountWithProps(flaw, { mode: 'create' });
 
     const titleField = subject
-      .findAllComponents(LabelEditable)
+      .findAllComponents(LabelDiv)
       .find(component => component.props().label === 'Title');
     expect(titleField?.exists()).toBe(true);
 
@@ -268,7 +268,7 @@ describe('flawForm', () => {
     expect(vm.errors.comment_zero).not.toBe(null);
 
     const titleField = subject
-      .findAllComponents(LabelEditable)
+      .findAllComponents(LabelDiv)
       .find(component => component.props().label === 'Title')
       ?.find('.is-invalid');
     expect(titleField?.exists()).toBe(true);